Problem: seclet must not be able to decrypt by itself
During normal operations, if seclet stores the key it is using to
encrypt the secrets, this is not secure as as long as one recovers that
key, the secrets aren't so secret anymore.

Solution: split the newly generated key into two parts

One part will be recorded on seclet, the other will be provided by the
user during authorization.

We're using Shamir's Secret Sharing Scheme for this (as implemented by
Daan Sprenkels at https://github.com/dsprenkels/sss) so it will be
possible in the future to change the configuration of the share rules
and implement variable n-out-of-m schemes.

Tests are still somewhat flaky and don't always work.